wpt_babel/content-security-policy/media-src/media-src-7_3_2.sub.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>
   Video track src attribute must match src list - negative test
  </title>
  <meta content="script-src * 'unsafe-inline'; media-src 'self';" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
 </head>
 <body>
  <h1>
   Video track src attribute must match src list - negative test
  </h1>
  <div id="log">
  </div>
  <script>
   "use strict";

var source_test = async_test("Disallowed track element onerror handler fires.");
var trackURL = location.protocol + "//{{domains[www]}}:{{ports[http][0]}}/media/foo.vtt";
var t_spv = async_test("Test that securitypolicyviolation events are fired");
var test_count = 1;
window.addEventListener("securitypolicyviolation", t_spv.step_func(function (e) {
  assert_equals(e.violatedDirective, "media-src");
  assert_equals(e.blockedURI, trackURL);
  if (--test_count <= 0) {
    t_spv.done();
  }
}));
function media_loaded(t) {
  t.step(function () {
    assert_unreached("Disllowed track source loaded.");
  });
  t.done();
}
function media_error_handler(t) {
  t.done();
}
  </script>
  <video controls="" crossorigin="" height="240" id="videoObject" onerror="media_error_handler(source_test)" width="320">
   <source id="audioSourceObject" src="/media/A4.ogv" type="audio/ogg"/>
   <track default="" id="trackObject" kind="subtitles" label="English" onerror="media_error_handler(source_test)" onload="media_loaded(source_test)" onloadeddata="media_loaded(source_test)" srclang="en"/>
  </video>
  <script>
   "use strict";

document.getElementById("trackObject").src = trackURL;
source_test.step(function () {
  source_test.set_status(source_test.FAIL);
});
setTimeout(function () {
  if (source_test.phase != source_test.phases.COMPLETE) {
    source_test.step(function () {
      assert_unreached("Onerror event never fired for track element.");
    });
    source_test.done();
  }
}, 2 * 1000);
  </script>
 </body>
</html>
